+++ pTk/mTk/generic/tkListbox.c
@@ -727,7 +727,7 @@ ListboxWidgetObjCmd(clientData, interp, objc, objv)
 	     * if they are selected
 	     */
 	    for (i = 0; i < listPtr->nElements; i++) {
-		if (Tcl_FindHashEntry(listPtr->selection, (char *)i) != NULL) {
+		if (Tcl_FindHashEntry(listPtr->selection, (char *) (size_t) i) != NULL) {
 		    sprintf(indexStringRep, "%d", i);
 		    Tcl_AppendElement(interp, indexStringRep);
 		}
@@ -1221,7 +1221,7 @@ ListboxSelectionSubCmd(interp, listPtr, objc, objv)
 	    }
 	    Tcl_SetObjResult(interp,
 		    Tcl_NewBooleanObj((Tcl_FindHashEntry(listPtr->selection,
-			    (char *)first) != NULL)));
+			    (char *) (size_t) first) != NULL)));
 	    result = TCL_OK;
 	    break;
 	}
@@ -1404,7 +1404,7 @@ ListboxGetItemAttributes(interp, listPtr, index)
     Tcl_HashEntry *entry;
     ItemAttr *attrs;
 
-    entry = Tcl_CreateHashEntry(listPtr->itemAttrTable, (char *)index, &new);
+    entry = Tcl_CreateHashEntry(listPtr->itemAttrTable, (char *) (size_t) index, &new);
     if (new) {
 	attrs = (ItemAttr *) ckalloc(sizeof(ItemAttr));
 	attrs->border = NULL;
@@ -1905,7 +1905,7 @@ DisplayListbox(clientData)
 	 * Lookup this item in the item attributes table, to see if it has
 	 * special foreground/background colors
 	 */
-	entry = Tcl_FindHashEntry(listPtr->itemAttrTable, (char *)i);
+	entry = Tcl_FindHashEntry(listPtr->itemAttrTable, (char *) (size_t) i);
 
 	/*
 	 * If the listbox is enabled, items may be drawn differently;
@@ -1913,7 +1913,7 @@ DisplayListbox(clientData)
 	 * or background colors.
 	 */
 	if (listPtr->state & STATE_NORMAL) {
-	    if (Tcl_FindHashEntry(listPtr->selection, (char *)i) != NULL) {
+	    if (Tcl_FindHashEntry(listPtr->selection, (char *) (size_t) i) != NULL) {
 		/* Selected items are drawn differently. */
 		gc = listPtr->selTextGC;
 		width = Tk_Width(tkwin) - 2*listPtr->inset;
@@ -1982,7 +1982,7 @@ DisplayListbox(clientData)
 		/* Draw bottom bevel */
 		if (i + 1 == listPtr->nElements ||
 			Tcl_FindHashEntry(listPtr->selection,
-				(char *)(i + 1)) == NULL ) {
+				(char *) (size_t) (i + 1)) == NULL ) {
 		    Tk_3DHorizontalBevel(tkwin, pixmap, selectedBg, x-left,
 			    y + listPtr->lineHeight - listPtr->selBorderWidth,
 			    width+left+right, listPtr->selBorderWidth, 0, 0, 0,
@@ -2413,13 +2413,13 @@ ListboxDeleteSubCmd(listPtr, first, last)
     widthChanged = 0;
     for (i = first; i <= last; i++) {
 	/* Remove selection information */
-	entry = Tcl_FindHashEntry(listPtr->selection, (char *)i);
+	entry = Tcl_FindHashEntry(listPtr->selection, (char *) (size_t) i);
 	if (entry != NULL) {
 	    listPtr->numSelected--;
 	    Tcl_DeleteHashEntry(entry);
 	}
 
-	entry = Tcl_FindHashEntry(listPtr->itemAttrTable, (char *)i);
+	entry = Tcl_FindHashEntry(listPtr->itemAttrTable, (char *) (size_t) i);
 	if (entry != NULL) {
 	    ckfree((char *)Tcl_GetHashValue(entry));
 	    Tcl_DeleteHashEntry(entry);
@@ -2993,7 +2993,7 @@ ListboxSelect(listPtr, first, last, select)
      * be, remove it.
      */
     for (i = first; i <= last; i++) {
-	entry = Tcl_FindHashEntry(listPtr->selection, (char *)i);
+	entry = Tcl_FindHashEntry(listPtr->selection, (char *) (size_t) i);
 	if (entry != NULL) {
 	    if (!select) {
 		Tcl_DeleteHashEntry(entry);
@@ -3005,7 +3005,7 @@ ListboxSelect(listPtr, first, last, select)
 	} else {
 	    if (select) {
 		entry = Tcl_CreateHashEntry(listPtr->selection,
-			(char *)i, &new);
+			(char *) (size_t) i, &new);
 		Tcl_SetHashValue(entry, (ClientData) NULL);
 		listPtr->numSelected++;
 		if (firstRedisplay < 0) {
@@ -3080,7 +3080,7 @@ ListboxFetchSelection(clientData, offset, buffer, maxBytes)
     needNewline = 0;
     Tcl_DStringInit(&selection);
     for (i = 0; i < listPtr->nElements; i++) {
-	entry = Tcl_FindHashEntry(listPtr->selection, (char *)i);
+	entry = Tcl_FindHashEntry(listPtr->selection, (char *) (size_t) i);
 	if (entry != NULL) {
 	    if (needNewline) {
 		Tcl_DStringAppend(&selection, "\n", 1);
@@ -3376,14 +3376,14 @@ ListboxListVarProc(clientData, interp, name1, name2, flags)
     if (listPtr->nElements < oldLength) {
 	for (i = listPtr->nElements; i < oldLength; i++) {
 	    /* Clean up selection */
-	    entry = Tcl_FindHashEntry(listPtr->selection, (char *)i);
+	    entry = Tcl_FindHashEntry(listPtr->selection, (char *) (size_t) i);
 	    if (entry != NULL) {
 		listPtr->numSelected--;
 		Tcl_DeleteHashEntry(entry);
 	    }
 
 	    /* Clean up attributes */
-	    entry = Tcl_FindHashEntry(listPtr->itemAttrTable, (char *)i);
+	    entry = Tcl_FindHashEntry(listPtr->itemAttrTable, (char *) (size_t) i);
 	    if (entry != NULL) {
 		ckfree((char *)Tcl_GetHashValue(entry));
 		Tcl_DeleteHashEntry(entry);
@@ -3452,21 +3452,21 @@ MigrateHashEntries(table, first, last, offset)
      * inside the for loops */
     if (offset > 0) {
 	for (i = last; i >= first; i--) {
-	    entry = Tcl_FindHashEntry(table, (char *)i);
+	    entry = Tcl_FindHashEntry(table, (char *) (size_t) i);
 	    if (entry != NULL) {
 		clientData = Tcl_GetHashValue(entry);
 		Tcl_DeleteHashEntry(entry);
-		entry = Tcl_CreateHashEntry(table, (char *)(i + offset), &new);
+		entry = Tcl_CreateHashEntry(table, (char *) (size_t) (i + offset), &new);
 		Tcl_SetHashValue(entry, clientData);
 	    }
 	}
     } else {
 	for (i = first; i <= last; i++) {
-	    entry = Tcl_FindHashEntry(table, (char *)i);
+	    entry = Tcl_FindHashEntry(table, (char *) (size_t) i);
 	    if (entry != NULL) {
 		clientData = Tcl_GetHashValue(entry);
 		Tcl_DeleteHashEntry(entry);
-		entry = Tcl_CreateHashEntry(table, (char *)(i + offset), &new);
+		entry = Tcl_CreateHashEntry(table, (char *) (size_t) (i + offset), &new);
 		Tcl_SetHashValue(entry, clientData);
 	    }
 	}
