+++ pTk/mTk/tclGeneric/tclHash.c
@@ -307,7 +307,7 @@ Tcl_FindHashEntry(tablePtr, key)
 	    index = hash & tablePtr->mask;
 	}
     } else {
-	hash = (unsigned int) key;
+	hash = (unsigned int) (size_t) key;
 	index = RANDOM_INDEX (tablePtr, hash);
     }
 
@@ -319,7 +319,7 @@ Tcl_FindHashEntry(tablePtr, key)
 	for (hPtr = tablePtr->buckets[index]; hPtr != NULL;
 	        hPtr = hPtr->nextPtr) {
 #if TCL_HASH_KEY_STORE_HASH
-	    if (hash != (unsigned int) hPtr->hash) {
+	    if (hash != (unsigned int) (size_t) hPtr->hash) {
 		continue;
 	    }
 #endif
@@ -331,7 +331,7 @@ Tcl_FindHashEntry(tablePtr, key)
 	for (hPtr = tablePtr->buckets[index]; hPtr != NULL;
 	        hPtr = hPtr->nextPtr) {
 #if TCL_HASH_KEY_STORE_HASH
-	    if (hash != (unsigned int) hPtr->hash) {
+	    if (hash != (unsigned int) (size_t) hPtr->hash) {
 		continue;
 	    }
 #endif
@@ -405,7 +405,7 @@ Tcl_CreateHashEntry(tablePtr, key, newPtr)
 	    index = hash & tablePtr->mask;
 	}
     } else {
-	hash = (unsigned int) key;
+	hash = (unsigned int) (size_t) key;
 	index = RANDOM_INDEX (tablePtr, hash);
     }
 
@@ -417,7 +417,7 @@ Tcl_CreateHashEntry(tablePtr, key, newPtr)
 	for (hPtr = tablePtr->buckets[index]; hPtr != NULL;
 	        hPtr = hPtr->nextPtr) {
 #if TCL_HASH_KEY_STORE_HASH
-	    if (hash != (unsigned int) hPtr->hash) {
+	    if (hash != (unsigned int) (size_t) hPtr->hash) {
 		continue;
 	    }
 #endif
@@ -430,7 +430,7 @@ Tcl_CreateHashEntry(tablePtr, key, newPtr)
 	for (hPtr = tablePtr->buckets[index]; hPtr != NULL;
 	        hPtr = hPtr->nextPtr) {
 #if TCL_HASH_KEY_STORE_HASH
-	    if (hash != (unsigned int) hPtr->hash) {
+	    if (hash != (unsigned int) (size_t) hPtr->hash) {
 		continue;
 	    }
 #endif
@@ -456,7 +456,7 @@ Tcl_CreateHashEntry(tablePtr, key, newPtr)
     hPtr->tablePtr = tablePtr;
 #if TCL_HASH_KEY_STORE_HASH
 #   if TCL_PRESERVE_BINARY_COMPATABILITY
-    hPtr->hash = (VOID *) hash;
+    hPtr->hash = (VOID *) (size_t) hash;
 #   else
     hPtr->hash = hash;
 #   endif
@@ -534,7 +534,7 @@ Tcl_DeleteHashEntry(entryPtr)
 	|| typePtr->flags & TCL_HASH_KEY_RANDOMIZE_HASH) {
 	index = RANDOM_INDEX (tablePtr, entryPtr->hash);
     } else {
-	index = ((unsigned int) entryPtr->hash) & tablePtr->mask;
+	index = ((unsigned int) (size_t) entryPtr->hash) & tablePtr->mask;
     }
 
     bucketPtr = &(tablePtr->buckets[index]);
@@ -1168,7 +1168,7 @@ RebuildTable(tablePtr)
 		|| typePtr->flags & TCL_HASH_KEY_RANDOMIZE_HASH) {
 		index = RANDOM_INDEX (tablePtr, hPtr->hash);
 	    } else {
-		index = ((unsigned int) hPtr->hash) & tablePtr->mask;
+		index = ((unsigned int) (size_t) hPtr->hash) & tablePtr->mask;
 	    }
 	    hPtr->nextPtr = tablePtr->buckets[index];
 	    tablePtr->buckets[index] = hPtr;
